+++ src/org/usfirst/frc/team5957/robot/RobotMap.java
@@ -1,4 +1,5 @@
 package org.usfirst.frc.team5957.robot;
+
 /**
  * The RobotMap is a mapping from the ports sensors and actuators are wired into
  * to a variable name. This provides flexibility changing wiring, makes checking
@@ -6,18 +7,18 @@ package org.usfirst.frc.team5957.robot;
  * floating around.
  */
 public class RobotMap {
-    // For example to map the left and right motors, you could define the
-    // following variables to use with your drivetrain subsystem.
-    // public static int leftMotor = 1;
-    // public static int rightMotor = 2;
-	public static final int leftFrontMotor  = 1;
+
+	// Motor Controller Ports
+	public static final int leftFrontMotor = 1;
 	public static final int rightFrontMotor = 3;
-	public static final int leftBackMotor   = 0;
-	public static final int rightBackMotor  = 2;
+	public static final int leftBackMotor = 0;
+	public static final int rightBackMotor = 2;
+	public static final int doorMotor = 4;
+
+	// Axis Camera Stream URL
 	public static final String cameraURL = "http://10.59.57.19/axis-cgi/mjpg/video.cgi";
-    
-    // If you are using multiple modules, make sure to define both the port
-    // number and the module. For example you with a rangefinder:
-    // public static int rangefinderPort = 1;
-    // public static int rangefinderModule = 1;
+
+	// DI/O Ports
+	public static final int doorLimit = 0;
+
 }
