+++ src/org/usfirst/frc/team5957/robot/commands/door/DoorHold.java
@@ -1,26 +1,26 @@
-package org.usfirst.frc.team5957.robot.commands;
+package org.usfirst.frc.team5957.robot.commands.door;
 
 import org.usfirst.frc.team5957.robot.Robot;
+
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
  *
  */
-public class DriveTrainBrake extends Command {
+public class DoorHold extends Command {
 
-    public DriveTrainBrake() {
-        // Use requires() here to declare subsystem dependencies
-    	super("DriveTrainBrake");
-    	requires(Robot.driveTrain);
+    public DoorHold() {
+        requires(Robot.door);
     }
 
     // Called just before this Command runs the first time
     protected void initialize() {
+    	Robot.door.stop();
     }
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-    	Robot.driveTrain.brake();
+    	Robot.door.stop();
     }
 
     // Make this return true when this Command no longer needs to run execute()
