+++ src/org/usfirst/frc/team5957/robot/commands/drivetrain/DriveTrainAimAtLift.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team5957.robot.commands;
+package org.usfirst.frc.team5957.robot.commands.drivetrain;
 
 import org.usfirst.frc.team5957.robot.Robot;
