+++ src/org/usfirst/frc/team5957/robot/commands/drivetrain/DriveTrainArcadeDrive.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team5957.robot.commands;
+package org.usfirst.frc.team5957.robot.commands.drivetrain;
 
 import org.usfirst.frc.team5957.robot.OI.ControllerType;
 import org.usfirst.frc.team5957.robot.Robot;
@@ -6,13 +6,14 @@ import org.usfirst.frc.team5957.robot.Robot;
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
- * A command that takes two joysticks' vertical values and drives the robot with them.
+ * A command that takes one joystick's vertical and horizontal values and
+ * drives the robot with them.
  */
-public class DriveTrainTankDrive extends Command {
+public class DriveTrainArcadeDrive extends Command {
 
-    public DriveTrainTankDrive() {
+    public DriveTrainArcadeDrive() {
         // Use requires() here to declare subsystem dependencies
-    	super("DriveTrainTankDrive");
+    	super("DriveTrainArcadeDrive");
     	requires(Robot.driveTrain);
     }
 
@@ -23,18 +24,18 @@ public class DriveTrainTankDrive extends Command {
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-    	double leftPower = 0;
-    	double rightPower = 0;
+    	double forward = 0;
+    	double turn = 0;
     	
     	if(Robot.oi.joystick == ControllerType.kGamepad) {
-    		leftPower = -Robot.oi.leftStick.getRawAxis(1); // left y
-    		rightPower = -Robot.oi.leftStick.getRawAxis(5); // right y
+    		forward = -Robot.oi.leftStick.getRawAxis(1); // left y
+    		turn    = -Robot.oi.leftStick.getRawAxis(0); // left y
     	} else if(Robot.oi.joystick == ControllerType.kFlightStick) {
-    		leftPower = -Robot.oi.leftStick.getY();
-    		rightPower = -Robot.oi.rightStick.getY();
+    		forward = -Robot.oi.leftStick.getY();
+    		turn = -Robot.oi.leftStick.getX();
     	}
-	
-    	Robot.driveTrain.tankDrive(leftPower, rightPower);
+    	
+    	Robot.driveTrain.arcadeDrive(forward, turn);
     }
 
     // Make this return true when this Command no longer needs to run execute()
