+++ src/org/usfirst/frc/team5957/robot/commands/drivetrain/DriveTrainBrake.java
@@ -1,35 +1,26 @@
-package org.usfirst.frc.team5957.robot.commands;
+package org.usfirst.frc.team5957.robot.commands.drivetrain;
 
 import org.usfirst.frc.team5957.robot.Robot;
-
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
  *
  */
-public class DriveTrainDriveAtPower extends Command {
+public class DriveTrainBrake extends Command {
 
-	double power = 0.25;
-	
-    public DriveTrainDriveAtPower() {
+    public DriveTrainBrake() {
         // Use requires() here to declare subsystem dependencies
-        // eg. requires(chassis);
+    	super("DriveTrainBrake");
     	requires(Robot.driveTrain);
-    }
-    
-    public DriveTrainDriveAtPower(double power) {
-    	this();
-    	this.power = power;
     }
 
     // Called just before this Command runs the first time
     protected void initialize() {
-    	Robot.driveTrain.brake();
     }
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-    	Robot.driveTrain.drive(power, 0);
+    	Robot.driveTrain.brake();
     }
 
     // Make this return true when this Command no longer needs to run execute()
@@ -39,12 +30,10 @@ public class DriveTrainDriveAtPower extends Command {
 
     // Called once after isFinished returns true
     protected void end() {
-    	Robot.driveTrain.brake();
     }
 
     // Called when another command which requires one or more of the same
     // subsystems is scheduled to run
     protected void interrupted() {
-    	Robot.driveTrain.brake();
     }
 }
