+++ src/org/usfirst/frc/team5957/robot/commands/drivetrain/DriveTrainDriveAtPower.java
@@ -1,61 +1,50 @@
-package org.usfirst.frc.team5957.robot.commands;
+package org.usfirst.frc.team5957.robot.commands.drivetrain;
 
 import org.usfirst.frc.team5957.robot.Robot;
 
 import edu.wpi.first.wpilibj.command.Command;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 /**
  *
  */
-public class DrivetrainTurn extends Command {
-	double angle = 90;
+public class DriveTrainDriveAtPower extends Command {
+
+	double power = 0.25;
 	
-    public DrivetrainTurn() {
+    public DriveTrainDriveAtPower() {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
     	requires(Robot.driveTrain);
     }
     
-    public DrivetrainTurn(double angle) {
+    public DriveTrainDriveAtPower(double power) {
     	this();
-    	this.angle = angle;
+    	this.power = power;
     }
-    
 
     // Called just before this Command runs the first time
     protected void initialize() {
-    	SmartDashboard.putString("Turning", ((Double) angle).toString().concat(" degrees"));
-    	Robot.oi.gyro.reset();
+    	Robot.driveTrain.brake();
     }
 
     // Called repeatedly when this Command is scheduled to run
     protected void execute() {
-    	if(angle < 0) {
-        	Robot.driveTrain.tankDrive(0.33, -0.33);
-    	} else {
-    		Robot.driveTrain.tankDrive(-0.33, 0.33);
-    	}
+    	Robot.driveTrain.drive(power, 0);
     }
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
-    	if(Math.abs(Robot.oi.gyro.getAngle()) > Math.abs(angle)) {
-    		return true;
-    	}
         return false;
     }
 
     // Called once after isFinished returns true
     protected void end() {
-    	SmartDashboard.putString("Turning", "Done");
     	Robot.driveTrain.brake();
     }
 
     // Called when another command which requires one or more of the same
     // subsystems is scheduled to run
     protected void interrupted() {
-    	SmartDashboard.putString("Turning", "Interrupted");
     	Robot.driveTrain.brake();
     }
 }
