+++ src/org/usfirst/frc/team5957/robot/commands/drivetrain/DrivetrainDriveAndTurn.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team5957.robot.commands;
+package org.usfirst.frc.team5957.robot.commands.drivetrain;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
